Add initial unit test harness
There is just a single test right now,
but it seems possible to extent the unit tests
in the future to cover the advertised UC API
and to run tun the tests on all supported backends.

Also add a .gitignore file to prevent compiled QML files
and vim swap file from getting committed. :)